Begin using state instead of raw element manipulation
