+++ src/components/tabs/console/Console.test.js
@@ -6,66 +6,4 @@ describe('Console', function() {
   before(function() {
     this.console = Console();
   });
-
-  describe('#setHistory()', function() {
-    it('should store history in reverse order', function() {
-      this.console.setHistory("'history1'");
-      this.console.setHistory("'history2'");
-      this.console.setHistory("'history3'");
-      assert.deepStrictEqual(this.console.getHistory(), ["'history3'", "'history2'", "'history1'"]);
-    });
-    it('should store the most recent 50 commands', function() {
-      for (let i = 1; i <= 60; i++) {
-        this.console.setHistory(`'history${i}'`);
-      }
-      const history = this.console.getHistory();
-      assert.strictEqual(history.length, 50);
-      assert.deepStrictEqual(history[0], "'history60'");
-      assert.deepStrictEqual(history[49], "'history11'");
-    });
-  });
-
-  describe('#getHistory()', function() {
-    it('should retrieve the history', function() {
-      this.console.setHistory(null, 0);
-      this.console.setHistory("'history'");
-      const history = this.console.getHistory();
-      assert.deepStrictEqual(history, ["'history'"]);
-      this.console.setHistory("'history2'");
-      const history2 = this.console.getHistory();
-      assert.deepStrictEqual(history2, ["'history2'", "'history'"]);
-    });
-  });
-
-  describe('#exec()', function() {
-    it('should output the input and the evaluated object reference', function() {
-      window.testString = 'value';
-      stdout.inspectSync((output) => {
-        this.console.exec('window.testString');
-        assert.deepStrictEqual(output, [
-          'window.testString\n',
-          `${window.testString}\n`
-        ]);
-      });
-    });
-    it('should assign variables', function() {
-      window.testString = 'old';
-      this.console.exec('window.testString = "new"');
-      assert.strictEqual(window.testString, 'new');
-    });
-    it('should not execute functions', function() {
-      let testVar = 'old';
-      window.testFn = () => (testVar = 'new');
-      this.console.exec('window.testFn()');
-      assert.strictEqual(testVar, 'old');
-    });
-    it('should catch errors and this.console.error it out', function() {
-      stderr.inspectSync((output) => {
-        assert.deepStrictEqual(output.length, 0, this.setupError);
-        const notAString = 123;
-        this.console.exec(notAString);
-        assert.deepStrictEqual(output.length, 1);
-      });
-    });
-  });
 });
