--- src/utilities/parseCommand.js
@@ -1,19 +1,0 @@
-const parseCommand = (command) => {
-  command = command.trim();
-  if (command.match(/^".*"$/) || command.match(/^'.*'$/)) {
-    return command.slice(1, -1);
-  }
-  const expressions = command.split(/\s*=\s*/);
-  const firstExpression = expressions.shift();
-  return firstExpression.replace(/(?=\[)/g, '.').split('.').reduce((object, memberString, idx, array) => {
-    const bracketMatch = memberString.match(/^\[([^\]]*)]$/);
-    const memberName = bracketMatch ? bracketMatch[1].replace(/^["']|["']$/g, '') : memberString;
-    if (expressions.length > 0 && idx === array.length - 1) {
-      // If there are things to the right of the equals sign, assign it to the left
-      (object || {})[memberName] = parseCommand(expressions.join('='));
-    }
-    return (object || {})[memberName];
-  }, window);
-};
-
-export default parseCommand;
